+++ laktory/models/resources/databricks/warehouse.py
@@ -1,7 +1,9 @@
+from typing import Any
 from typing import Literal
 from typing import Union
 
 from pydantic import Field
+from pydantic import model_validator
 
 from laktory.models.basemodel import BaseModel
 from laktory.models.resources.baseresource import ResourceLookup
@@ -52,9 +54,23 @@ class WarehouseLookup(ResourceLookup):
     ----------
     id:
         The ID of the SQL warehouse.
+    name:
+        Name of the SQL warehouse. Name of the SQL warehouse to search (case-sensitive).
+        Argument only supported by Terraform IaC backend.
     """
 
-    id: str = Field(serialization_alias="id")
+    id: str = Field(serialization_alias="id", default=None)
+    name: str = None
+
+    @model_validator(mode="after")
+    def at_least_one(self) -> Any:
+        if self.id is None and self.name is None:
+            raise ValueError("At least `id` or `name` must be set.")
+
+        if not (self.id is None or self.name is None):
+            raise ValueError("Only one of `id` or `name` must be set.")
+
+        return self
 
 
 class Warehouse(BaseModel, PulumiResource, TerraformResource):